git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@6038 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CMS3DMeshFileLoader.cpp
@@ -192,8 +192,8 @@ bool CMS3DMeshFileLoader::load(io::IReadFile* file)
 	// read whole file
 
 	u8* buffer = new u8[fileSize];
-	s32 read = file->read(buffer, fileSize);
-	if (read != fileSize)
+	size_t read = file->read(buffer, fileSize);
+	if (read != (size_t)fileSize)
 	{
 		delete [] buffer;
 		os::Printer::log("Could not read full file. Loading failed", file->getFileName(), ELL_ERROR);
